fix(types): add type annotations and fix mypy errors across codebase
- Add return type annotations to __init__ and __post_init__ methods
- Add type: ignore comments for complex generics and overloads
- Fix variable shadowing in loop iterations
- Add proper union type annotations for conditional assignments
- Fix callable type mismatches in hook registrations
- Update dict type annotations to use Any where needed

Affected modules:
- control_plane/watchdog.py - return types and stats dict typing
- streaming/resilience.py - __init__ return type
- gateway/enterprise/proxy.py - hook variable naming
- gateway/enterprise/audit_interceptor.py - redacted list typing
- caching/decorators.py - overload implementation typing
- agents/telemetry.py - BaseException error handling
- observability/otel.py - processor union type
- persistence/migrations/consolidate.py - dict access typing
- debate/event_bus.py - __post_init__ return type
- db/profiling.py - sqlite execute parameter typing

+++ aragora/skills/builtin/file_analysis.py
@@ -125,7 +125,7 @@ class FileMetrics:
         }
 
     @staticmethod
-    def _format_size(size_bytes: int) -> str:
+    def _format_size(size_bytes: float) -> str:
         """Format byte size to human readable."""
         for unit in ["B", "KB", "MB", "GB"]:
             if size_bytes < 1024:
